#22 tiny refactoring and folder movement
