+++ src/middlewares/index.ts
@@ -1,4 +1,4 @@
 import { ValidationConfig } from './validation-config.interface';
-import { validationConfigPipe } from './validation-config-pipe';
+import { validationConfigPipe } from './validation-config.pipe';
 
 export { ValidationConfig, validationConfigPipe };
