--- src/validation-body-pipe/index.ts
@@ -1,4 +1,0 @@
-import { validationBodyPipe } from './validation-body-pipe';
-import { DefaultBodyError } from './errors/default-body.error';
-
-export { validationBodyPipe, DefaultBodyError };
