--- src/validation-query-pipe/index.ts
@@ -1,4 +1,0 @@
-import { validationQueryPipe } from './validation-query-pipe';
-import { DefaultQueryError } from './errors/default-query.error';
-
-export { validationQueryPipe, DefaultQueryError };
