+++ src/validators/index.ts
@@ -1,0 +1,7 @@
+import { DefaultBodyError } from './body/errors/default-body.error';
+import { validationBodyPipe } from './body/validation-body.pipe';
+
+import { DefaultQueryError } from './query/errors/default-query.error';
+import { validationQueryPipe } from './query/validation-query.pipe';
+
+export { DefaultBodyError, validationBodyPipe, DefaultQueryError, validationQueryPipe };
