+++ src/validators/query/errors/default-query.error.ts
@@ -1,5 +1,5 @@
 import { StatusCodes } from 'http-status-codes';
-import { BaseError, ErrorField } from '../../common/errors';
+import { BaseError, ErrorField } from '../../../common/errors';
 
 export class DefaultQueryError extends BaseError {
     constructor(public readonly fields: ErrorField[]) {
